refactor: organize landing page code for better structure
