+++ app/Http/Controllers/LandingPageController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class LandingPageController extends Controller
+{
+    public function index(){
+        return view('landing_page', [
+            'title' => 'landingPage',
+        ]);
+    }
+}
